logging: Fix timestamp 32bit integer overflow
We are specifically requesting wall time for the logging output, but
using 32 bits for micro-seconds does overflow, leaving us with an
arbitrary integer.

+++ src/wayland-client.c
@@ -1557,7 +1557,7 @@ queue_event(struct wl_display *display, int len)
 	const struct wl_message *message;
 	struct wl_event_queue *queue;
 	struct timespec tp;
-	unsigned int time;
+	uint64_t time;
 	int num_zombie_fds;
 
 	wl_connection_copy(display->connection, p, sizeof p);
@@ -1577,11 +1577,11 @@ queue_event(struct wl_display *display, int len)
 
 		if (debug_client) {
 			clock_gettime(CLOCK_REALTIME, &tp);
-			time = (tp.tv_sec * 1000000L) + (tp.tv_nsec / 1000);
+			time = (tp.tv_sec * 1000000LL) + (tp.tv_nsec / 1000);
 
-			fprintf(stderr, "[%7u.%03u] discarded [%s]#%d.[event %d]"
+			fprintf(stderr, "[%7" PRIu64 ".%03u] discarded [%s]#%d.[event %d]"
 				"(%d fd, %d byte)\n",
-				time / 1000, time % 1000,
+				time / 1000, (unsigned int) (time % 1000),
 				zombie ? "zombie" : "unknown",
 				id, opcode,
 				num_zombie_fds, size);
